Improve multiplication overflow situation
By first multiplying the number of milliseconds with the clock frequency
(which is often a large number, e.g. 8000000) before dividing it
further, it was easy to run into an overflow case starting at 537 ms, as
537 * 8000000 = 4296000000 which is greater than 2^32 = 4294967296.
By only multiplying the intended delay with the already divided clock
frequency, we allow for larger delays. This only shifts the problem
though, the maximum delay is now 1073741 ms assuming a clock frequency
of 8 MHz. The only way to improve this further is to switch to u64.